// arg-web-frontend/src/types/accounts-payable.types.ts
export interface CompanyMaintenanceData {
  companyNo: number;
  companyName: string;
  companyApGlNo: number;
  companyBankGlNo: number;
  companyDiscountsGlNo: number;
  companyIntercoGlNo: number;
  companyNextPjJrnlNo: number;
  companyNextCdJrnlNo: number;
  companyNextCheckNo: number;
  companyNextEntryNo: number;
  companyNextVoucherNo: number;
  companyPreEdChks: "Y" | "N";
  companyJobCostAct: "Y" | "N";
  companyRetentionGlNo: number;
  companyPoActive: "Y" | "N";
  companyEmployeeExpenseGlNo: number;
  companyNextEeJrnlNo: number;
  companyFiller: string;
}
// ...
export interface GlMasterValidationResult {
  success: boolean;
  message: string;
  description?: string;
  isDeleted?: boolean;
}